Operator Development addcmul #886
